requirements.txt
flask==3.0.0
plexapi==4.15.10
requests==2.31.0
PyYAML==6.0.1
